moving parsePassword to helper package
